back/requirements.txt
Django
django-debug-toolbar
django-anymail[sendinblue]
dj_database_url
django-cors-headers
django-extensions
django-ipware
django-request-id
djangorestframework
djangorestframework-gis
djangorestframework-simplejwt
django-filter
factory_boy
flake8
freezegun
getconf
html2text
ipdb
ipython
jinja2
psycopg2-binary
requests
responses
sqlparse
whitenoise
gunicorn
